requirements.txt
flask==3.0.0
google-generativeai==0.3.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
python-dateutil==2.8.2
python-dotenv==1.0.1
SpeechRecognition==3.10.1
gTTS==2.5.1
dateparser==1.2.0
pytz==2024.1 